Rename: domain to model
