+++ .gitignore
@@ -3,6 +3,7 @@ target/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
+voucher_data.txt
 
 ### STS ###
 .apt_generated
