+++ src/main/java/org/prgrms/kdt/model/FixedAmountVoucher.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt.domain;
+package org.prgrms.kdt.model;
 
 import java.io.Serializable;
 import java.util.UUID;
