+++ src/main/java/org/prgrms/kdt/repository/MemOrderRepository.java
@@ -1,7 +1,7 @@
 package org.prgrms.kdt.repository;
 
 
-import org.prgrms.kdt.domain.Order;
+import org.prgrms.kdt.model.Order;
 import org.springframework.stereotype.Repository;
 
 import java.util.*;
