+++ src/main/java/org/prgrms/kdt/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package org.prgrms.kdt.repository;
 
-import org.prgrms.kdt.domain.Order;
+import org.prgrms.kdt.model.Order;
 
 public interface OrderRepository {
     Order insert(Order order);
